Fixed Error message for Amount in Transfer/Add ,Fixed radio button in add and details Transfer

+++ VirtualWallet/VirtualWallet.DTO/ExchangeDto/CreateExcahngeDto.cs
@@ -16,7 +16,7 @@ namespace VirtualWallet.Dto.CreateExcahngeDto
         public string To { get; set; }
 
         [Required]
-        [Range(0.01, double.MaxValue, ErrorMessage = "The {0} must be greater than 0.")]
+        [Range(0.01, (double)decimal.MaxValue, ErrorMessage = "The {0} must be greater than 0.")]
         public decimal Amount { get; set; }
 
     }
